requirements.txt
pyyaml
mock
pyinstaller
requests
joblib
pynacl
